+++ src/components/Tipseen/Tipseen.jsx
@@ -1,5 +1,5 @@
-import { DialogPosition } from "../../constants/sizes";
-import { useRef, forwardRef, useMemo, Fragment } from "react";
+import { DialogPosition } from "../../constants/positions";
+import { forwardRef, Fragment, useMemo, useRef } from "react";
 import PropTypes from "prop-types";
 import cx from "classnames";
 import useMergeRefs from "../../hooks/useMergeRefs";
@@ -8,7 +8,7 @@ import Button from "../../components/Button/Button";
 import { BEMClass } from "../../helpers/bem-helper";
 import Icon from "../../components/Icon/Icon";
 import CloseSmall from "../../components/Icon/Icons/components/CloseSmall";
-import { JustifyType, AnimationType } from "../../constants/dialog";
+import { AnimationType, JustifyType } from "../../constants/dialog";
 import TipseenTitle from "./TipseenTitle";
 import { TIPSEEN_CLOSE_BUTTON_ARIA_LABEL } from "./TipseenConstants";
 import styles from "./Tipseen.module.scss";
