+++ src/constants/sizes.ts
@@ -7,18 +7,3 @@ export const BASE_SIZES = {
 export const BASE_SIZES_WITH_NONE = { NONE: "none", ...BASE_SIZES };
 
 export const SIZES = { XXS: "xxs", XS: "xs", ...BASE_SIZES } as const;
-
-export const DialogPosition = {
-  LEFT: "left",
-  LEFT_START: "left-start",
-  LEFT_END: "left-end",
-  RIGHT: "right",
-  RIGHT_START: "right-start",
-  RIGHT_END: "right-end",
-  TOP: "top",
-  TOP_START: "top-start",
-  TOP_END: "top-end",
-  BOTTOM: "bottom",
-  BOTTOM_START: "bottom-start",
-  BOTTOM_END: "bottom-end"
-} as const;
